my-fastapi-app/pyproject.toml
[project]
name = "my-fastapi-app"
version = "0.1.0"
description = ""
authors = [
    {name = "nansystem",email = "sato171786@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.110.0,<0.111.0)",
    "uvicorn (>=0.27.1,<0.28.0)",
    "pydantic (>=2.7.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "cloud-sql-python-connector (>=1.16.0,<2.0.0)",
    "cryptography (>=44.0.0,<45.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"
pre-commit = "^4.1.0"

[tool.ruff]
target-version = "py311"
include = ["*.py", "*.pyi", "**/pyproject.toml"]
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "migrations",
]
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.ruff.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# B008 FastAPIのDependsの警告を無視
[tool.ruff.lint]
ignore = ["E203", "E266", "E501", "E722", "E741", "W291", "W293", "B008"]
